<==== Mercurial data ====>
Branch: webadmin

+++ wixsetup/maven/filter-resources/wix-gen.py
@@ -16,8 +16,10 @@ for wxs in ('dbsync', 'help', 'vox', 'bg'):
 		
 if os.path.exists(join(bin_source_dir, '${product.name} Launcher.exe')):
     os.unlink(join(bin_source_dir, '${product.name} Launcher.exe'))
-shutil.copy2(join(bin_source_dir, 'applauncher.exe'), join(bin_source_dir, '${product.name} Launcher.exe'))
+if os.path.exists(join(bin_source_dir, 'applauncher.exe')):
+    shutil.copy2(join(bin_source_dir, 'applauncher.exe'), join(bin_source_dir, '${product.name} Launcher.exe'))
     
 if os.path.exists(join(bin_source_dir, '${product.name}.exe')):
     os.unlink(join(bin_source_dir, '${product.name}.exe'))          
-shutil.copy2(join(bin_source_dir, 'client.exe'), join(bin_source_dir, '${product.name}.exe'))
+if os.path.exists(join(bin_source_dir, 'client.exe')):
+    shutil.copy2(join(bin_source_dir, 'client.exe'), join(bin_source_dir, '${product.name}.exe'))
